test: split payment tests to parallelize on CI
test: split to parallelize

chore: remove

chore

ci
