+++ specs/e2e/shipments.spec.ts
@@ -578,157 +578,6 @@ test.describe("with do no track SKU", () => {
   })
 })
 
-test.describe("no shipping zone with cart url", () => {
-  test.use({
-    defaultParams: {
-      order: "with-items",
-      lineItemsAttributes: [
-        { sku_code: "CANVASAU000000FFFFFF1824", quantity: 1 },
-      ],
-      orderAttributes: {
-        cart_url: faker.internet.url(),
-        customer_email: customerEmail,
-      },
-      addresses: {
-        billingAddress: {
-          ...usAddress,
-          billing_info: faker.random.alphaNumeric(11),
-        },
-        sameShippingAddress: true,
-      },
-    },
-  })
-
-  test("no shipping method to select", async ({ checkoutPage }) => {
-    await checkoutPage.checkOrderSummary("Order Summary")
-    await checkoutPage.checkStep("Shipping", "open")
-    await checkoutPage.checkButton({ type: "Shipping", status: "not_present" })
-    const element = checkoutPage.page.locator(
-      "text=The entered destination is outside our shipping zone."
-    )
-    await expect(element).toHaveCount(1)
-  })
-})
-
-test.describe("one item shippable and one item out of stock with cartUrl", () => {
-  test.use({
-    defaultParams: {
-      order: "with-items",
-      lineItemsAttributes: [
-        { sku_code: "NOSTOCK", quantity: 1, inventory: 0 },
-        { sku_code: "CANVASAU000000FFFFFF1824", quantity: 1 },
-      ],
-      orderAttributes: {
-        cart_url: faker.internet.url(),
-        customer_email: customerEmail,
-      },
-      addresses: {
-        billingAddress: euAddress,
-        sameShippingAddress: true,
-      },
-    },
-  })
-
-  test("no shipping method to select", async ({ checkoutPage }) => {
-    await checkoutPage.checkOrderSummary("Order Summary")
-    await checkoutPage.checkStep("Shipping", "open")
-    await checkoutPage.checkButton({
-      type: "Shipping",
-      status: "not_present",
-    })
-    const element = checkoutPage.page.locator(
-      "text=An item in your order is no longer available. Click here to edit your cart."
-    )
-    await expect(element).toHaveCount(1)
-  })
-})
-
-test.describe("only out of stock", () => {
-  test.use({
-    defaultParams: {
-      order: "with-items",
-      lineItemsAttributes: [{ sku_code: "NOSTOCK", quantity: 1, inventory: 0 }],
-      orderAttributes: {
-        cart_url: faker.internet.url(),
-        customer_email: customerEmail,
-      },
-      addresses: {
-        billingAddress: euAddress,
-        sameShippingAddress: true,
-      },
-    },
-  })
-
-  test("no shipping method to select", async ({ checkoutPage }) => {
-    await checkoutPage.checkOrderSummary("Order Summary")
-    await checkoutPage.checkStep("Shipping", "open")
-    await checkoutPage.checkButton({ type: "Shipping", status: "not_present" })
-    const element = checkoutPage.page.locator(
-      "text=An item in your order is no longer available. Click here to edit your cart."
-    )
-    await expect(element).toHaveCount(1)
-  })
-})
-
-test.describe("partial out of stock no cart_url", () => {
-  test.use({
-    defaultParams: {
-      order: "with-items",
-      lineItemsAttributes: [{ sku_code: "NOSTOCK", quantity: 2, inventory: 1 }],
-      orderAttributes: {
-        customer_email: customerEmail,
-      },
-      addresses: {
-        billingAddress: euAddress,
-        sameShippingAddress: true,
-      },
-    },
-  })
-
-  test("no shipping method to select", async ({ checkoutPage }) => {
-    await checkoutPage.checkOrderSummary("Order Summary")
-    await checkoutPage.checkStep("Shipping", "open")
-    await checkoutPage.checkButton({ type: "Shipping", status: "not_present" })
-    const element = checkoutPage.page.locator(
-      "text=An item in your order is no longer available."
-    )
-    await expect(element).toHaveCount(1)
-  })
-})
-
-test.describe("no shipping zone and one out of stock", () => {
-  test.use({
-    defaultParams: {
-      order: "with-items",
-      lineItemsAttributes: [
-        { sku_code: "NOSTOCK", quantity: 1, inventory: 0 },
-        { sku_code: "CANVASAU000000FFFFFF1824", quantity: 1 },
-      ],
-      orderAttributes: {
-        cart_url: faker.internet.url(),
-        customer_email: customerEmail,
-      },
-      addresses: {
-        billingAddress: {
-          ...usAddress,
-          billing_info: faker.random.alphaNumeric(11),
-        },
-        sameShippingAddress: true,
-      },
-    },
-  })
-
-  test("no shipping method to select", async ({ checkoutPage }) => {
-    await checkoutPage.checkOrderSummary("Order Summary")
-    await checkoutPage.checkStep("Shipping", "open")
-    await checkoutPage.checkButton({ type: "Shipping", status: "not_present" })
-    const element = checkoutPage.page.locator(
-      "text=An item in your order is no longer available. Click here to edit your cart."
-    )
-    await expect(element).toHaveCount(1)
-  })
-})
-
 test.describe("ship from primary", () => {
   test.use({
     defaultParams: {
